+++ auth-module/src/main/java/com/inhabas/api/auth/domain/token/jwtUtils/JwtTokenProvider.java
@@ -1,5 +1,6 @@
 package com.inhabas.api.auth.domain.token.jwtUtils;
 
+import com.inhabas.api.auth.domain.token.securityFilter.TokenAuthenticationProcessingFilter;
 import com.inhabas.api.auth.domain.token.TokenDto;
 import com.inhabas.api.auth.domain.token.TokenProvider;
 import com.inhabas.api.auth.domain.token.jwtUtils.refreshToken.RefreshToken;
