+++ auth-module/src/main/java/com/inhabas/api/auth/domain/token/securityFilter/InvalidJwtTokenHandler.java
@@ -1,4 +1,4 @@
-package com.inhabas.api.auth.domain.token.jwtUtils;
+package com.inhabas.api.auth.domain.token.securityFilter;
 
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
