+++ resource-server/src/main/java/com/inhabas/api/domain/member/MemberAuthorityProvider.java
@@ -3,6 +3,8 @@ package com.inhabas.api.domain.member;
 import com.inhabas.api.auth.domain.oauth2.socialAccount.type.UID;
 import com.inhabas.api.auth.domain.oauth2.userAuthorityProvider.UserAuthorityProvider;
 import com.inhabas.api.auth.domain.oauth2.userInfo.OAuth2UserInfo;
+import com.inhabas.api.domain.member.security.socialAccount.MemberSocialAccount;
+import com.inhabas.api.domain.member.security.socialAccount.MemberSocialAccountRepository;
 import com.inhabas.api.domain.member.type.wrapper.Role;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
